+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+
+## 0.2.0
+* adds support for trinkets 
+* optimize core events code (run better)
+* on player disconnect backpack is saving
+
 ## 0.1.2
 * optimize mixins amount
 * repair NBT problem after crafting backpack